refactor(commands): move command catalog into src/shared
Relocate the platform-agnostic command metadata catalog from
packages/extension/src/commands/catalog.ts to src/shared/commands/catalog.ts
so the file path matches the import boundary already enforced by tsconfig
paths. Both VS Code and desktop already share the catalog via the
@commands/* alias; this move makes the shared ownership explicit by placing
it alongside the existing shared dispatcher in src/shared/commands.
