+++ src/shared/commands/catalog.ts
